Moved schema stuff from views to schema, return timestamp correctly
